Plug in new pipeline - all tests pass
