+++ Tests/AppTests/PipelineTests.swift
@@ -123,7 +123,7 @@ class PipelineTests: AppTestCase {
         }
 
         // MUT - second stage
-        try ingest(client: app.client, database: app.db, limit: 10).wait()
+        try ingest(application: app, database: app.db, limit: 10).wait()
 
         do { // validate
             let packages = try Package.query(on: app.db).sort(\.$url).all().wait()
@@ -156,7 +156,7 @@ class PipelineTests: AppTestCase {
         }
 
         // MUT - ingest again
-        try ingest(client: app.client, database: app.db, limit: 10).wait()
+        try ingest(application: app, database: app.db, limit: 10).wait()
 
         do {  // validate - only new package moves to .ingestion stage
             let packages = try Package.query(on: app.db).sort(\.$url).all().wait()
@@ -181,7 +181,7 @@ class PipelineTests: AppTestCase {
         Current.date = { Date().addingTimeInterval(Constants.reIngestionDeadtime) }
 
         // MUT - ingest yet again
-        try ingest(client: app.client, database: app.db, limit: 10).wait()
+        try ingest(application: app, database: app.db, limit: 10).wait()
 
         do {  // validate - now all three packages should have been updated
             let packages = try Package.query(on: app.db).sort(\.$url).all().wait()
